requirements.txt
tenacity~=8.2.3
requests~=2.31.0
psycopg2~=2.9.9
pandas~=2.2.1
matplotlib~=3.8.3
plotly~=5.20.0
numpy~=1.26.4
scikit-learn~=1.4.1.post1
scipy~=1.12.0
gensim~=4.3.2
fastapi==0.110.0
uvicorn==0.29.0
websockets
wsproto
uvicorn[standard]
python-dotenv~=1.0.1
fastapi-users[sqlalchemy]
slowapi
typing_inspect
fastapi_restful
asyncpg
alembic
httpx
datetime